updates to log target
