+++ ucsd_cloud_cli/logs/__init__.py
@@ -1,6 +1,7 @@
 import click
-from .target import target
+from .target import cli as target
+from .accounts import cli as account
 from .. import cf_data_dir
 
 
-logs = click.CommandCollection(sources=[target])
+logs = click.CommandCollection(sources=[target, account])
